Fix sticky scrolling bug in draggable piles.
Previous to this commit there was a bug where if you clicked the
scrollbar of a pile a certain way it would cause the pile to stick
to the mouse and follow the cursor around. This was fixed by
adding a handle to the top of the pile to target for the sortable
drag action.