Implement reset button on scores. #298
Was going to do the click-magic, but raty-widget already has 'cancel',
and 'cancelTrue' options that use a button. Further, the user might not
guess that clicking the first star again will turn it off. That's something
you found intuitive, but others might not. The best solution is to give the
use a visual instruction, such as a cancel button that is self-evident.

The cancel graphic that comes with raty is not pretty. We will just need
our future designer to make us a better graphic.